Fix building and renames
