+++ bungeecord/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 }
 
 dependencies {
+    api(projects.common)
     compileOnly("net.md-5:bungeecord-api:1.19-R0.1-SNAPSHOT")
 }
