+++ bungeecord/src/main/java/dev/kejona/geyserblockplatform/bungeecord/GeyserBlockPlatformBungee.java
@@ -2,9 +2,9 @@ package dev.kejona.geyserblockplatform.bungeecord;
 
 import dev.kejona.geyserblockplatform.common.BlockResult;
 import dev.kejona.geyserblockplatform.common.Permissions;
-import dev.kejona.geyserblockplatform.common.platformchecker.GeyserBedrockPlatformChecker;
-import dev.kejona.geyserblockplatform.common.platformchecker.BedrockPlatformChecker;
-import dev.kejona.geyserblockplatform.common.platformchecker.FloodgateBedrockPlatformChecker;
+import dev.kejona.geyserblockplatform.common.bedrock.BaseApiHandler;
+import dev.kejona.geyserblockplatform.common.bedrock.BedrockHandler;
+import dev.kejona.geyserblockplatform.common.bedrock.FloodgateHandler;
 import dev.kejona.geyserblockplatform.common.config.Config;
 import dev.kejona.geyserblockplatform.common.config.ConfigLoader;
 import net.md_5.bungee.api.ProxyServer;
@@ -20,16 +20,16 @@ import net.md_5.bungee.event.EventPriority;
 import java.io.IOException;
 
 public final class GeyserBlockPlatformBungee extends Plugin implements Listener {
-    private BedrockPlatformChecker handler;
+    private BedrockHandler handler;
     private Config config;
 
     @Override
     public void onEnable() {
         PluginManager pluginManager = getProxy().getPluginManager();
         if (pluginManager.getPlugin("floodgate") != null) {
-            handler = new FloodgateBedrockPlatformChecker();
+            handler = new FloodgateHandler();
         } else if (pluginManager.getPlugin("Geyser-Bungeecord") != null) {
-            handler = new GeyserBedrockPlatformChecker();
+            handler = new BaseApiHandler();
         } else {
             getLogger().warning("There is no Geyser or Floodgate plugin detected! Disabling...");
             onDisable();
@@ -53,7 +53,7 @@ public final class GeyserBlockPlatformBungee extends Plugin implements Listener 
         }
 
         BlockResult result = config.computeResult(player.getUniqueId(), handler);
-        if (!result.isAllowed()) {
+        if (!result.allowed()) {
             BlockResult.Denied deniedResult = (BlockResult.Denied) result;
             player.disconnect(TextComponent.fromLegacyText(deniedResult.message()));
         } else {
