+++ common/src/main/java/dev/kejona/geyserblockplatform/common/bedrock/BedrockHandler.java
@@ -1,0 +1,8 @@
+package dev.kejona.geyserblockplatform.common.bedrock;
+
+import java.util.UUID;
+
+public interface BedrockHandler {
+    boolean isBedrockPlayer(UUID uuid);
+    Profile profile(UUID uuid);
+}
