--- common/src/main/java/dev/kejona/geyserblockplatform/common/lombok.config
@@ -1,1 +1,0 @@
-lombok.accessors.fluent=true
