--- common/src/main/java/dev/kejona/geyserblockplatform/common/platformchecker/BedrockPlatformChecker.java
@@ -1,10 +1,0 @@
-package dev.kejona.geyserblockplatform.common.platformchecker;
-
-import dev.kejona.geyserblockplatform.common.Profile;
-
-import java.util.UUID;
-
-public interface BedrockPlatformChecker {
-    boolean isBedrockPlayer(UUID uuid);
-    Profile profile(UUID uuid);
-}
