+++ spigot/src/main/java/dev/kejona/geyserblockplatform/spigot/GeyserBlockPlatformSpigot.java
@@ -2,9 +2,9 @@ package dev.kejona.geyserblockplatform.spigot;
 
 import dev.kejona.geyserblockplatform.common.BlockResult;
 import dev.kejona.geyserblockplatform.common.Permissions;
-import dev.kejona.geyserblockplatform.common.platformchecker.GeyserBedrockPlatformChecker;
-import dev.kejona.geyserblockplatform.common.platformchecker.BedrockPlatformChecker;
-import dev.kejona.geyserblockplatform.common.platformchecker.FloodgateBedrockPlatformChecker;
+import dev.kejona.geyserblockplatform.common.bedrock.BaseApiHandler;
+import dev.kejona.geyserblockplatform.common.bedrock.BedrockHandler;
+import dev.kejona.geyserblockplatform.common.bedrock.FloodgateHandler;
 import dev.kejona.geyserblockplatform.common.config.Config;
 import dev.kejona.geyserblockplatform.common.config.ConfigLoader;
 import org.bukkit.Bukkit;
@@ -19,16 +19,16 @@ import org.bukkit.plugin.java.JavaPlugin;
 import java.io.IOException;
 
 public final class GeyserBlockPlatformSpigot extends JavaPlugin implements Listener {
-    private BedrockPlatformChecker handler;
+    private BedrockHandler handler;
     private Config config;
 
     @Override
     public void onEnable() {
         PluginManager pluginManager = Bukkit.getPluginManager();
         if (pluginManager.getPlugin("floodgate") != null) {
-            handler = new FloodgateBedrockPlatformChecker();
+            handler = new FloodgateHandler();
         } else if (pluginManager.getPlugin("Geyser-Spigot") != null) {
-            handler = new GeyserBedrockPlatformChecker();
+            handler = new BaseApiHandler();
         } else {
             getLogger().warning("There is no Geyser or Floodgate plugin detected! Disabling...");
             Bukkit.getPluginManager().disablePlugin(this);
@@ -56,7 +56,7 @@ public final class GeyserBlockPlatformSpigot extends JavaPlugin implements Liste
 
         BlockResult result = config.computeResult(player.getUniqueId(), handler);
 
-        if (!result.isAllowed()) {
+        if (!result.allowed()) {
             BlockResult.Denied deniedResult = (BlockResult.Denied) result;
             event.setResult(PlayerLoginEvent.Result.KICK_OTHER);
             event.setKickMessage(deniedResult.message());
